drivers/smmstore: Support 64-bit MMIO addresses
Currently the SMMSTOREv2 only supports MMIO ROM below 4GiB. As the
space below 4GiB is typically limited on x86, add support for extended
MMIO ROM windows as found on recent hardware.

Allow the SMMSTOREv2 to be memory mapped above 4GiB by adding a new
field to the coreboot table called 'mmap_addr'. The users outside
of coreboot must check the size field of the coreboot struct to
determine if mmap_addr is supported. When it is it holds the
64-bit physical address to the MMIO ROM window.
The old 'mmap_addr' was renamed to 'mmap_addr_deprecated' to indicate
that it should not be used any more.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/87114
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/include/smmstore.h
@@ -69,7 +69,7 @@ struct smmstore_params_init {
 struct smmstore_params_info {
 	uint32_t num_blocks;
 	uint32_t block_size;
-	uint32_t mmap_addr;
+	uint64_t mmap_addr;
 } __packed;
 
 /*
